# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Việt phục Remix API",
    "description": "Backend REST API phục vụ nền tảng phối đồ Việt phục Remix (VietStylist)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ai/try-on": {
      "post": {
        "tags": [
          "AI Virtual Try-On"
        ],
        "summary": "Create Try On Job",
        "description": "Tạo tác vụ thử đồ Việt phục trên ảnh người dùng bằng AI (F05).\nTác vụ được đưa vào hàng đợi xử lý bất đồng bộ bởi Worker.",
        "operationId": "create_try_on_job_api_ai_try_on_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTryOnJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TryOnJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/jobs/{job_id}": {
      "get": {
        "tags": [
          "AI Virtual Try-On"
        ],
        "summary": "Get Job Status",
        "description": "Truy vấn trạng thái tác vụ AI (queued -> running -> succeeded/failed) và ảnh kết quả.",
        "operationId": "get_job_status_api_ai_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TryOnJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/solution-form": {
      "get": {
        "tags": [
          "Solution Form F12"
        ],
        "summary": "Get Solution Form",
        "description": "Lấy nội dung form trình bày giải pháp của đội (F12).\nNếu chưa có, tự động khởi tạo bản nháp mẫu với đầy đủ các mục nghiên cứu và văn hóa.",
        "operationId": "get_solution_form_api_solution_form_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionFormResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Solution Form F12"
        ],
        "summary": "Update Solution Form",
        "description": "Cập nhật nội dung form giải pháp, phát hiện xung đột khi mở nhiều tab (F12).",
        "operationId": "update_solution_form_api_solution_form_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolutionFormRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionFormResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTryOnJobRequest": {
        "properties": {
          "user_photo_url": {
            "type": "string",
            "title": "User Photo Url"
          },
          "outfit_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outfit Version Id"
          },
          "outfit_snapshot": {
            "$ref": "#/components/schemas/OutfitSnapshot-Input"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "user_photo_url",
          "outfit_snapshot"
        ],
        "title": "CreateTryOnJobRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemTransform": {
        "properties": {
          "dx": {
            "type": "number",
            "title": "Dx",
            "default": 0
          },
          "dy": {
            "type": "number",
            "title": "Dy",
            "default": 0
          },
          "scale": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Scale",
            "default": 1
          },
          "rotation": {
            "type": "number",
            "title": "Rotation",
            "default": 0
          }
        },
        "type": "object",
        "title": "ItemTransform"
      },
...
      "LookbookRef": {
        "properties": {
          "lookbook_id": {
            "type": "string",
            "title": "Lookbook Id"
          },
          "lookbook_title": {
            "type": "string",
            "title": "Lookbook Title"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          }
        },
        "type": "object",
        "required": [
          "lookbook_id",
          "lookbook_title"
        ],
        "title": "LookbookRef"
      },
...
      "OutfitSnapshot-Input": {
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "title": "Schemaversion",
            "default": 1
          },
          "avatarId": {
            "type": "string",
            "title": "Avatarid",
            "default": "avatar_nam_chuan"
          },
          "poseId": {
            "type": "string",
            "title": "Poseid",
            "default": "front_01"
          },
          "occasionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasionid",
            "default": "ky_yeu"
          },
          "styleMode": {
            "type": "string",
            "title": "Stylemode",
            "default": "traditional"
          },
          "overlapDirection": {
            "type": "string",
            "title": "Overlapdirection",
            "default": "right_over_left"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotItem"
            },
            "type": "array",
            "title": "Items",
            "default": []
          },
          "lockedSlots": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lockedslots"
          },
          "backgroundTheme": {
            "type": "string",
            "enum": [
              "white",
              "dopaper"
            ],
            "title": "Backgroundtheme",
            "default": "white"
          },
          "aspectRatio": {
            "type": "string",
            "enum": [
              "1:1",
              "9:16"
            ],
            "title": "Aspectratio",
            "default": "9:16"
          }
        },
        "type": "object",
        "title": "OutfitSnapshot"
      },
...
      "SnapshotItem": {
        "properties": {
          "slot": {
            "type": "string",
            "title": "Slot"
          },
          "itemId": {
            "type": "string",
            "title": "Itemid"
          },
          "variantId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variantid"
          },
          "assetVersion": {
            "type": "integer",
            "title": "Assetversion",
            "default": 1
          },
          "colorOptionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coloroptionid"
          },
          "colorHex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colorhex"
          },
          "transform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemTransform"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "slot",
          "itemId"
        ],
        "title": "SnapshotItem"
      },
      "SolutionFormResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "proposed_solution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Solution"
          },
          "cultural_safeguards": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultural Safeguards"
          },
          "lookbook_references": {
            "items": {
              "$ref": "#/components/schemas/LookbookRef"
            },
            "type": "array",
            "title": "Lookbook References",
            "default": []
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "team_name",
          "product_name",
          "revision",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SolutionFormResponse"
      },
...
      "TryOnJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": "try_on"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "outfit_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outfit Version Id"
          },
          "result_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Image Url"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "input_hash",
          "created_at",
          "updated_at"
        ],
        "title": "TryOnJobResponse"
      },
...
      "UpdateSolutionFormRequest": {
        "properties": {
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "proposed_solution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Solution"
          },
          "cultural_safeguards": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultural Safeguards"
          },
          "lookbook_references": {
            "items": {
              "$ref": "#/components/schemas/LookbookRef"
            },
            "type": "array",
            "title": "Lookbook References",
            "default": []
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "team_name",
          "product_name",
          "revision"
        ],
        "title": "UpdateSolutionFormRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}